Remove white background behind account icons
The white background color causes a weird outline between the borders
and images in the mailbox list. Removing these declarations makes them
look as expected.

Additionally the material-ui package was pinned to 0.15.0-alpha.2,
because the one it currently resolves to is 0.15.0-beta.2, which has
dependency issues with react and react plugins/components.